reformat admin domain finder
