+++ kifi-backend/modules/shoebox/app/views/admin/admin.scala.html
@@ -76,9 +76,6 @@
             <ul class="dropdown-menu" role="menu">
               <li><a href="@com.keepit.controllers.admin.routes.SliderAdminController.getVersionForm">Keeper Versions</a>
               <li><a href="@com.keepit.controllers.admin.routes.SliderAdminController.getPatterns">Automatic Engagement</a>
-              <li><a href="@com.keepit.controllers.admin.routes.SliderAdminController.getDomainTags">Domain Tags</a>
-              <li><a href="@com.keepit.controllers.admin.routes.SliderAdminController.getDomainOverrides">Domain Overrides</a>
-              <li><a href="@com.keepit.controllers.admin.routes.SliderAdminController.searchDomain">Find Domain</a>
             </ul>
 
                <li class="dropdown">
@@ -134,6 +131,9 @@
                         <li><a href="@com.keepit.controllers.admin.routes.UrlController.urlRenormalizeConsole">URL Renormalization Console</a>
                         <li><a href="@com.keepit.controllers.admin.routes.UrlController.renormalizationView(0)">Renormalization</a>
                         <li><a href="@com.keepit.controllers.admin.routes.UrlController.getPatterns()">URL Pattern Rules</a>
+                        <li><a href="@com.keepit.controllers.admin.routes.UrlController.searchDomain">Find Domain</a>
+                        <li><a href="@com.keepit.controllers.admin.routes.UrlController.getDomainTags">Domain Tags</a>
+                        <li><a href="@com.keepit.controllers.admin.routes.UrlController.getDomainOverrides">Domain Overrides</a>
                         <li><a href="@com.keepit.controllers.admin.routes.ScraperAdminController.getProxies">Proxies</a>
                     </ul>
                 </li>
