+++ kifi-backend/modules/shoebox/app/views/admin/domainFind.scala.html
@@ -1,7 +1,7 @@
 @()(implicit request: com.keepit.common.controller.UserRequest[_])
 
 @admin("Find Domain", stylesheets = List("admin_slider")) {
-    <form action="@com.keepit.controllers.admin.routes.SliderAdminController.findDomainByHostname" method="POST" class="form-inline">
+    <form action="@com.keepit.controllers.admin.routes.UrlController.findDomainByHostname" method="POST" class="form-inline">
         <div class="form-group">
             <label for="hostname" class="control-label">Domain hostname</label>
             <input type="text" name="hostname" class="form-control" placeholder="Hostname">
