+++ kifi-backend/modules/shoebox/app/views/admin/domainTags.scala.html
@@ -1,7 +1,7 @@
 @(tags: Seq[com.keepit.classify.DomainTag])(implicit request: com.keepit.common.controller.UserRequest[_])
 
 @admin("Domain Tags", stylesheets = List("admin_slider")) {
-<form class="domain-tags" action="@com.keepit.controllers.admin.routes.SliderAdminController.saveDomainTags" method="POST">
+<form class="domain-tags" action="@com.keepit.controllers.admin.routes.UrlController.saveDomainTags" method="POST">
   <table class="table table-condensed table-striped">
     <thead>
       <tr>
